services/inpaint-api/pyproject.toml
[project]
name = "inpaint-api"
version = "0.1.0"
description = "FastAPI watermark removal service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "numpy>=2.0.2,<3.0.0",
  "opencv-python-headless>=4.10.0.84,<5.0.0.0",
  "pillow>=11.0.0,<12.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "python-multipart>=0.0.12,<1.0.0",
  "uvicorn>=0.32.0,<1.0.0"
]

[dependency-groups]
dev = [
  "httpx>=0.27.2,<1.0.0",
  "pytest>=8.3.3,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
